[feature] BIG Commit - Better UI, VHS recording speed, noise seed and overall code cleanups
This commit changes the overall code and UI. Most of the changes based on the pull request
on the original ntscQT by valadaptive. This includes:

- New feature: Noise seed (which allows you to set a seed for all video effects)
- New feature: VHS recording speed switch (SP, LP and EP)
- A better UI, overall
- Code cleanups

You can find more about these changes on his original pull request:
https://github.com/JargeZ/ntscqt/pull/94

Other change, this time outside of the pull, is the logo which is now in
the topleft corner of the UI.

- Logo on topleft UI

Thanks to valadaptive for the changes.